# application/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "基于区块链的钢材供应链融资解决方案api文档",
        "title": "基于区块链的钢材供应链融资解决方案",
        "contact": {
            "name": "togettoyou",
            "email": "144240096@qq.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/QueryFinance": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "查询所有",
                "parameters": [
                    {
                        "description": "Finance",
                        "name": "Finance",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.ProductByNameQueryRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/QueryFinanceByOrderID": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "按订单号查询",
                "parameters": [
                    {
                        "description": "Finance",
                        "name": "Finance",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.FinanceByOrderIDQueryRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/QueryFinanceByStatus": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "按融资状态查询",
                "parameters": [
                    {
                        "description": "Finance",
                        "name": "Finance",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.FinanceByStatusQueryRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/QueryFinanceBySupplier": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "查询所有",
                "parameters": [
                    {
                        "description": "Finance",
                        "name": "Finance",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.FinanceBySupplierQueryRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/StoreFinance": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "存储到数据库",
                "parameters": [
                    {
                        "description": "Finance",
                        "name": "Finance",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.FinanceRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/UpdateFinanceStatus": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "更新融资状态",
                "parameters": [
                    {
                        "description": "Finance",
                        "name": "Finance",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.UpdateStatusRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "app.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "v1.FinanceByOrderIDQueryRequestBody": {
            "type": "object",
            "properties": {
                "OrderID": {
                    "description": "对应订单号",
                    "type": "string"
                }
            }
        },
        "v1.FinanceByStatusQueryRequestBody": {
            "type": "object",
            "properties": {
                "Status": {
                    "description": "融资状态",
                    "type": "string"
                }
            }
        },
        "v1.FinanceBySupplierQueryRequestBody": {
            "type": "object",
            "properties": {
                "Supplier": {
                    "description": "申请人",
                    "type": "string"
                }
            }
        },
        "v1.FinanceRequestBody": {
            "type": "object",
            "properties": {
                "Amount": {
                    "description": "申请金额",
                    "type": "string"
                },
                "Desc": {
                    "description": "申请描述",
                    "type": "string"
                },
                "Feedback": {
                    "description": "融资反馈",
                    "type": "string"
                },
                "OrderID": {
                    "description": "对应订单号",
                    "type": "string"
                },
                "Period": {
                    "description": "申请日期",
                    "type": "string"
                },
                "Status": {
                    "description": "融资状态",
                    "type": "string"
                },
                "Supplier": {
                    "description": "申请人",
                    "type": "string"
                }
            }
        },
...
        "v1.ProductByNameQueryRequestBody": {
            "type": "object",
            "properties": {
                "GoodsID": {
                    "description": "货品ID",
                    "type": "string"
                }
            }
        },
...
        "v1.UpdateStatusRequestBody": {
            "type": "object",
            "properties": {
                "OrderID": {
                    "description": "对应订单号",
                    "type": "string"
                },
                "Period": {
                    "description": "申请日期",
                    "type": "string"
                },
                "Status": {
                    "description": "融资状态",
                    "type": "string"
                }
            }
        },
...
    }
}